requirements.txt
numpy
fastapi
fasthtml
httpx>=0.23.0
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=9.4.0
flask>=2.2.0
requests>=2.28.0